new application documents serializer for caseworker added
