+++ api/applications/caseworker/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 
+from api.applications.models import ApplicationDocument
 from api.staticdata.statuses.enums import CaseStatusEnum
 
 
@@ -7,3 +8,9 @@ class ApplicationChangeStatusSerializer(serializers.Serializer):
 
     status = serializers.ChoiceField(choices=CaseStatusEnum.all())
     note = serializers.CharField(required=False, allow_blank=True, allow_null=True, max_length=2000)
+
+
+class ApplicationDocumentSerializer(serializers.Serializer):
+    class Meta:
+        model = ApplicationDocument
+        fields = "__all__"
