+++ api/applications/caseworker/views/applications.py
@@ -1,40 +1,23 @@
-from django.core.exceptions import ObjectDoesNotExist
-from django.http import Http404, JsonResponse
+from django.http import JsonResponse
 from django.db import transaction
 from rest_framework.generics import GenericAPIView
 from rest_framework import status
 
 from api.applications.caseworker.permissions import CaseStatusCaseworkerChangeable
-from api.applications.caseworker.serializers import ApplicationChangeStatusSerializer
+from api.applications.caseworker.serializers import ApplicationChangeStatusSerializer, ApplicationDocumentSerializer
+from api.applications.caseworker.views.mixins import CaseworkerApplicationMixin
 from api.applications.helpers import get_application_view_serializer
-from api.applications.libraries.get_applications import get_application
-from api.core.exceptions import NotFoundError
-from api.core.authentication import GovAuthentication
+from api.applications.models import ApplicationDocument
 from api.core.permissions import CaseInCaseworkerOperableStatus
 from api.staticdata.statuses.libraries.get_case_status import get_case_status_by_status
 
 
-class ApplicationChangeStatus(GenericAPIView):
-    authentication_classes = (GovAuthentication,)
+class ApplicationChangeStatus(CaseworkerApplicationMixin, GenericAPIView):
     permission_classes = [
         CaseInCaseworkerOperableStatus,
         CaseStatusCaseworkerChangeable,
     ]
     serializer_class = ApplicationChangeStatusSerializer
-
-    def setup(self, request, *args, **kwargs):
-        super().setup(request, *args, **kwargs)
-        try:
-            self.application = get_application(self.kwargs["pk"])
-        except (ObjectDoesNotExist, NotFoundError):
-            raise Http404()
-
-    def get_object(self):
-        self.check_object_permissions(self.request, self.application)
-        return self.application
-
-    def get_case(self):
-        return self.application
 
     @transaction.atomic
     def post(self, request, pk):
@@ -49,3 +32,11 @@ class ApplicationChangeStatus(GenericAPIView):
         ).data
 
         return JsonResponse(data=response_data, status=status.HTTP_200_OK)
+
+
+class ApplicationDocumentView(CaseworkerApplicationMixin, GenericAPIView):
+    serializer_class = ApplicationDocumentSerializer
+    lookup_url_kwarg = "pk"
+
+    def get_queryset(self):
+        return ApplicationDocument.objects.filter(application_id=self.application.pk)
